+++ skbio/diversity/alpha/__init__.py
@@ -1,15 +1,25 @@
-"""Alpha diversity measures (:mod:`skbio.diversity.alpha`)
+r"""Alpha diversity measures (:mod:`skbio.diversity.alpha`)
 =======================================================
 
 .. currentmodule:: skbio.diversity.alpha
 
-This package provides implementations of alpha diversity measures, including
-measures of richness, dominance, and evenness. Some functions generate
-confidence intervals (CIs). These functions have the suffix ``_ci``.
+This package provides implementations of various alpha diversity [1]_ metrics,
+including measures of richness, diversity, evenness, dominance, coverage, and
+other categories of metrics.
+
+Some functions generate confidence intervals (CIs). These functions have
+the suffix ``_ci``.
 
 
 Richness metrics
 ----------------
+
+**Richness** [2]_ measures the number of species (taxa) in a community.
+
+Due to incomplete sampling, the number of observed species (``sobs``) in a
+sample is usually lower than the true number of species in the community.
+Metrics have been proposed to estimate the latter based on the distribution
+of observed species in the sample.
 
 .. autosummary::
    :toctree:
@@ -21,6 +31,7 @@ Richness metrics
    faith_pd
    margalef
    menhinick
+   michaelis_menten_fit
    observed_features
    observed_otus
    osd
@@ -28,8 +39,35 @@ Richness metrics
    sobs
 
 
+Diversity metrics
+-----------------
+
+**Diversity** [3]_ measures the number and relative abundances of species
+(taxa) in a community. It combines richness and evenness.
+
+Some diversity metrics describe the effective number of species (a.k.a., true
+diversity) -- the number of equally-abundant species that produce the same
+diversity measurement.
+
+.. autosummary::
+   :toctree:
+
+   brillouin_d
+   enspie
+   fisher_alpha
+   inv_simpson
+   kempton_taylor_q
+   phydiv
+   shannon
+   simpson
+
+
 Evenness metrics
 ----------------
+
+**Evenness** [4]_ (or equitability) measures the closeness of species (taxa) in a
+community in terms of abundance (number of individuals within the species). The
+calculation of evenness involves the relative abundances of species.
 
 .. autosummary::
    :toctree:
@@ -40,22 +78,14 @@ Evenness metrics
    simpson_e
 
 
-Diversity metrics
------------------
-
-.. autosummary::
-   :toctree:
-
-   brillouin_d
-   enspie
-   inv_simpson
-   phydiv
-   shannon
-   simpson
-
-
 Dominance metrics
 -----------------
+
+**Dominance** [5]_ (or concentration) measures the degree that one or a few
+most abundant species (taxa) represent the great majority of a community. It
+can be considered as a measure of community unevenness.
+
+It should be noted that higher dominance corresponds to lower biodiversity.
 
 .. autosummary::
    :toctree:
@@ -68,20 +98,37 @@ Dominance metrics
    strong
 
 
-Miscellaneous
--------------
+Coverage metrics
+----------------
+
+**Coverage** [6]_ measures the proportion of individuals of a community that
+have been observed (or unobserved) in a sample. It describes the completeness
+of sampling.
 
 .. autosummary::
    :toctree:
 
    esty_ci
-   fisher_alpha
    goods_coverage
-   kempton_taylor_q
    lladser_ci
    lladser_pe
-   michaelis_menten_fit
    robbins
+
+
+References
+----------
+.. [1] https://en.wikipedia.org/wiki/Alpha_diversity
+
+.. [2] https://en.wikipedia.org/wiki/Species_richness
+
+.. [3] https://en.wikipedia.org/wiki/Species_diversity
+
+.. [4] https://en.wikipedia.org/wiki/Species_evenness
+
+.. [5] https://en.wikipedia.org/wiki/Dominance_(ecology)
+
+.. [6] Good, I. J. (1953). The population frequencies of species and the
+   estimation of population parameters. Biometrika, 40(3-4), 237-264.
 
 """  # noqa: D205, D415